refactored for MVP pattern
