+++ ASE_android/app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MapsActivity" />
+        <activity android:name=".MapsScreen.GoogleMaps.MapsActivity" />
         <activity android:name=".LoginActivity" />
         <activity android:name=".SignUpActivity"></activity>
     </application>
