+++ ASE_android/app/src/main/java/sussex/android/ase_android/LoginActivity.java
@@ -14,6 +14,8 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 
+import sussex.android.ase_android.MapsScreen.GoogleMaps.MapsActivity;
+
 public class LoginActivity extends AppCompatActivity implements
         View.OnClickListener{
     private FirebaseAuth mAuth;
