+++ ase_android/app/src/main/java/sussex/android/ase_android/BasePresenter.java
@@ -1,0 +1,5 @@
+package sussex.android.ase_android;
+
+public interface BasePresenter {
+
+}
