+++ ase_android/app/src/main/java/sussex/android/ase_android/MapsScreen/BottomSheet/BottomSheetContract.java
@@ -1,0 +1,16 @@
+package sussex.android.ase_android.MapsScreen.BottomSheet;
+
+
+import sussex.android.ase_android.BasePresenter;
+
+public interface BottomSheetContract {
+        interface View {
+                int getPeekHeightPx();
+                void collapseBottomSheet();
+                void hideBottomSheet();
+        }
+
+        interface Presenter extends BasePresenter {
+
+        }
+}
