+++ ase_android/app/src/main/java/sussex/android/ase_android/MapsScreen/GoogleMaps/MapsContract.java
@@ -1,0 +1,14 @@
+package sussex.android.ase_android.MapsScreen.GoogleMaps;
+
+import android.app.Activity;
+
+import sussex.android.ase_android.BasePresenter;
+
+public interface MapsContract {
+    interface View {
+        Activity getActivity();
+    }
+
+    interface Presenter extends BasePresenter {
+    }
+}
