+++ ase_android/app/src/main/res/layout/activity_maps.xml
@@ -17,7 +17,7 @@
             android:name="com.google.android.gms.maps.SupportMapFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".MapsActivity" />
+            tools:context=".MapsScreen.GoogleMaps.MapsActivity" />
     </android.support.v4.widget.DrawerLayout>
     <include layout="@layout/layout_pull_up" />
 </android.support.design.widget.CoordinatorLayout>
